requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
SQLAlchemy>=2.0
psycopg2-binary>=2.9
pydantic>=2.5
pydantic-settings>=2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
alembic>=1.13
pytest>=8.0
httpx>=0.27
slowapi>=0.1.9
python-multipart>=0.0.9
email-validator>=2.1
asyncpg>=0.29.0